requirements.txt
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-db-connection-pool==1.2.4
Django==5.2.1
drf-spectacular==0.26.2
django-filter==23.5
drf-nested-routers==0.93.4
gevent==25.5.1
gunicorn==20.1.0
python-decouple==3.8
psycopg2-binary==2.9.10
PyJWT==2.6.0
python-dateutil==2.8.2
whitenoise==6.9.0
